refactor: project viewed event handler moved to viewed package
