+++ entities-viewings-collector/src/main/scala/io/renku/entities/viewings/collector/projects/viewed/SubscriptionFactory.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.entities.viewings.collector.projects
+package io.renku.entities.viewings.collector.projects.viewed
 
 import cats.effect.Async
 import cats.syntax.all._
